requirements.txt
postal
chardet
orjson
sqlglot[rs]
pandas
python-magic
charset_normalizer